Changed tubeValidation.js
